+ constants environment based
